Refactored Team Tab Panel and Settings page
