+++ Client/src/Components/TabPanels/ProfileSettings/ProfilePanel.jsx
@@ -19,7 +19,6 @@ import Button from "../../Button";
  * and deleting account.
  *
  * @returns {JSX.Element}
- *
  */
 
 const ProfilePanel = () => {
